Add article about solving vulnerability
Resolving vulnerabilities in React Native applications is crucial to
ensure the security and stability of the codebase. We will explore the
tools and techniques that can aid in resolving vulnerabilities
efficiently. We will highlight the importance of utilizing Yarn
Resolutions in ``package.json``, which allows you to enforce specific
versions of dependencies, overriding the versions specified in the
yarn.lock file. Additionally, we will discuss the benefits of using
Patch-package, a tool that enables you to make modifications to
dependencies without directly modifying the source code in cooperation
with postinstall. However patch-package should be considered as a last
resort in the fixing crashes/issues or vulnerability resolution process,
when other methods such as enforcing specific versions through Yarn
Resolutions or managing dependency are not sufficient to address the
vulnerabilities. The first step should be to search library repositories
and look for vulnerability fixes.